requirements.txt
tabulate
pycryptodome